+++ python/code/s3.py
@@ -1,0 +1,81 @@
+# The MIT License (MIT)
+# =====================
+#
+# Copyright © 2020 Azavea
+#
+# Permission is hereby granted, free of charge, to any person
+# obtaining a copy of this software and associated documentation
+# files (the “Software”), to deal in the Software without
+# restriction, including without limitation the rights to use,
+# copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following
+# conditions:
+#
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+# OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+# HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+# WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+# OTHER DEALINGS IN THE SOFTWARE.
+
+
+def parse_s3_url(url: str) -> Tuple[str, str]:
+    """Given an S3 URI, return the bucket and prefix
+
+    Arguments:
+        url {str} -- The S3 URI to parse
+
+    Returns:
+        Tuple[str, str] -- The bucket and prefix as a tuple
+    """
+    parsed = urlparse(url, allow_fragments=False)
+    return (parsed.netloc, parsed.path.lstrip('/'))
+
+
+def get_matching_s3_keys(bucket: str,
+                         prefix: str = '',
+                         suffix: str = '') -> Generator[str, None, None]:
+    """Generate all of the keys in a bucket with the given prefix and suffix
+
+    See https://alexwlchan.net/2017/07/listing-s3-keys/
+
+    Arguments:
+        bucket {str} -- The S3 bucket
+
+    Keyword Arguments:
+        prefix {str} -- The prefix to filter by (default: {''})
+        suffix {str} -- The suffix to filter by (default: {''})
+
+    Returns:
+        Generator[str, None, None] -- The list of keys
+    """
+    s3 = boto3.client('s3')
+    kwargs = {'Bucket': bucket}
+
+    # If the prefix is a single string (not a tuple of strings), we can
+    # do the filtering directly in the S3 API.
+    if isinstance(prefix, str):
+        kwargs['Prefix'] = prefix
+
+    while True:
+        # The S3 API response is a large blob of metadata.
+        # 'Contents' contains information about the listed objects.
+        resp = s3.list_objects_v2(**kwargs)
+        for obj in resp['Contents']:
+            key = obj['Key']
+            if key.startswith(prefix) and key.endswith(suffix):
+                yield key
+
+        # The S3 API is paginated, returning up to 1000 keys at a time.
+        # Pass the continuation token into the next response, until we
+        # reach the final page (when this field is missing).
+        try:
+            kwargs['ContinuationToken'] = resp['NextContinuationToken']
+        except KeyError:
+            break
